professional freelancers updated to be under serviceProviders entity
