+++ src/app.module.ts
@@ -8,7 +8,6 @@ import { UsersModule } from './modules/users/users.module';
 import { SmsModule } from './shared/sms/sms.module';
 import { UploadModule } from './shared/upload/upload.module';
 import { Employer } from './modules/users/entities/employer.entity';
-import { Freelancer } from './modules/users/entities/freelancer.entity';
 import { ServiceProvider } from './modules/users/entities/serviceProvider.entity';
 import { PropertyOwner } from './modules/users/entities/propertyOwner.entity';
 import { PropertyRenter } from './modules/users/entities/propertyRenter.entity';
@@ -33,7 +32,6 @@ import { User } from './modules/users/entities/users.entity';
         entities: [
           User,
           Employer,
-          Freelancer,
           ServiceProvider,
           PropertyOwner,
           PropertyRenter,
