+++ src/modules/auth/auth.module.ts
@@ -9,7 +9,6 @@ import { MailerConfigModule } from 'src/shared/mailer/mailer.module';
 import { SmsModule } from 'src/shared/sms/sms.module';
 import { UploadModule } from 'src/shared/upload/upload.module';
 import { Employer } from '../users/entities/employer.entity';
-import { Freelancer } from '../users/entities/freelancer.entity';
 import { ServiceProvider } from '../users/entities/serviceProvider.entity';
 import { PropertyOwner } from '../users/entities/propertyOwner.entity';
 import { PropertyRenter } from '../users/entities/propertyRenter.entity';
@@ -22,7 +21,6 @@ import { User } from '../users/entities/users.entity';
     TypeOrmModule.forFeature([
       User,
       Employer,
-      Freelancer,
       ServiceProvider,
       PropertyOwner,
       PropertyRenter,
